# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  gymtrack-backend_internal_domain_models.TrainerProfile:
    properties:
      bio:
        type: string
      hourlyRate:
        type: number
      isAvailableForNewClients:
        type: boolean
      languages:
        items:
          type: string
        type: array
      location:
        type: string
      profilePhotoUrl:
        type: string
      yearsOfExperience:
        type: integer
    type: object
  gymtrack-backend_internal_domain_models.TrainerReview:
    properties:
      athleteId:
        type: string
      comment:
        type: string
      createdAt:
        type: string
      rating:
        maximum: 5
        minimum: 1
        type: integer
      reviewId:
        type: string
      trainerId:
        type: string
      type:
        description: Always "review"
        type: string
      updatedAt:
        type: string
    required:
    - rating
    type: object
  gymtrack-backend_internal_domain_models.TrainerWithProfile:
    properties:
      averageRating:
        type: number
      createdAt:
        type: string
      email:
        type: string
      passwordHash:
        type: string
      profile:
        $ref: '#/definitions/gymtrack-backend_internal_domain_models.TrainerProfile'
      reviewCount:
        type: integer
      role:
        allOf:
        - $ref: '#/definitions/gymtrack-backend_internal_domain_models.UserRole'
        enum:
        - trainer
        - athlete
      type:
        description: Always "user"
        type: string
      updatedAt:
        type: string
      userId:
        type: string
    required:
    - email
    - role
    type: object
...
  gymtrack-backend_internal_domain_models.UserRole:
    enum:
    - trainer
    - athlete
    type: string
    x-enum-varnames:
    - RoleTrainer
    - RoleAthlete
...
  internal_api_handlers.CreateReviewRequest:
    description: Request body for creating a review for a trainer.
    properties:
      comment:
        type: string
      rating:
        maximum: 5
        minimum: 1
        type: integer
    required:
    - rating
    type: object
...
  internal_api_handlers.UpdateReviewRequest:
    description: Request body for updating an existing review.
    properties:
      comment:
        type: string
      rating:
        maximum: 5
        minimum: 1
        type: integer
    required:
    - rating
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Fitness tracking API for personal trainers and athletes
  title: GymTrack API
  version: "1.0"
paths:
...
  /reviews/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes an existing review. Only the author can delete their review.
      parameters:
      - description: Review ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Review deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: User not authenticated
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Only the author can delete the review
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Review not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to delete review
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a review
      tags:
      - Reviews
    put:
      consumes:
      - application/json
      description: Updates an existing review. Only the author can update their review.
      parameters:
      - description: Review ID
        in: path
        name: id
        required: true
        type: string
      - description: Update review request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_handlers.UpdateReviewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Review updated successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request body or validation failed
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: User not authenticated
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Only the author can update the review
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Review not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to update review
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a review
      tags:
      - Reviews
  /trainers:
    get:
      consumes:
      - application/json
      description: Retrieves a list of trainers with optional filtering by specialization,
        location, minimum rating, and availability status. Supports pagination.
      parameters:
      - description: Filter by specialization (e.g., 'yoga', 'strength')
        in: query
        name: specialization
        type: string
      - description: Filter by location
        in: query
        name: location
        type: string
      - description: Minimum rating filter (0-5)
        in: query
        name: minRating
        type: number
      - description: Filter by availability for new clients (true/false)
        in: query
        name: availableForNewClients
        type: boolean
      - description: 'Number of results per page (default: 20)'
        in: query
        name: limit
        type: integer
      - description: 'Number of results to skip (default: 0)'
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved trainers" "{\"trainers\":[],\"total\":0,\"limit\":20,\"offset\":0}
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error" "{\"error\":\"error message\"}
          schema:
            additionalProperties: true
            type: object
      summary: Get all trainers with optional filtering
      tags:
      - Trainers
  /trainers/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves detailed profile information for a specific trainer by
        their ID
      parameters:
      - description: Trainer ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved trainer
          schema:
            $ref: '#/definitions/gymtrack-backend_internal_domain_models.TrainerWithProfile'
        "404":
          description: Trainer not found" "{\"error\":\"trainer not found\"}
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error" "{\"error\":\"error message\"}
          schema:
            additionalProperties: true
            type: object
      summary: Get trainer by ID
      tags:
      - Trainers
...
  /trainers/{id}/reviews:
    get:
      description: Retrieve all reviews for a specific trainer. This endpoint is publicly
        accessible.
      parameters:
      - description: Trainer ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'Reviews retrieved successfully" {"reviews": []models.TrainerReview}'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to retrieve reviews
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get reviews for a trainer
      tags:
      - Reviews
    post:
      consumes:
      - application/json
      description: Allows an athlete to create a review for a specific trainer. Requires
        athlete role.
      parameters:
      - description: Trainer ID
        in: path
        name: id
        required: true
        type: string
      - description: Create review request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_handlers.CreateReviewRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Review created successfully
          schema:
            $ref: '#/definitions/gymtrack-backend_internal_domain_models.TrainerReview'
        "400":
          description: Invalid request body or validation failed
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: User not authenticated
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Only athletes can create reviews
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to create review
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a review for a trainer
      tags:
      - Reviews
...
  /trainers/me/profile:
    get:
      consumes:
      - application/json
      description: Retrieves the profile information for the currently authenticated
        trainer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved trainer profile
          schema:
            $ref: '#/definitions/gymtrack-backend_internal_domain_models.TrainerWithProfile'
        "401":
          description: Unauthorized" "{\"error\":\"unauthorized\"}
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden - only trainers can access" "{\"error\":\"Only trainers
            can access profile\"}
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Profile not found" "{\"error\":\"trainer profile not found\"}
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error" "{\"error\":\"error message\"}
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get current trainer's profile
      tags:
      - Trainers
    put:
      consumes:
      - application/json
      description: Updates the profile information for the currently authenticated
        trainer
      parameters:
      - description: Profile update data
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/gymtrack-backend_internal_domain_models.TrainerProfile'
      produces:
      - application/json
      responses:
        "200":
          description: Profile updated successfully" "{\"message\":\"profile updated
            successfully\"}
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request data" "{\"error\":\"error message\"}
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized" "{\"error\":\"unauthorized\"}
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error" "{\"error\":\"error message\"}
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update current trainer's profile
      tags:
      - Trainers
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"